Search across all datasets
fix(search): modify search to work across all datasets. Additionally,
add a SEARCH_KEY property to each object to deal with the fact that HASH_KEY
can change. The search indicies need a key to use to identify an object,
the neccesity of changing the HASH_KEY property to handle user definitions
of "sameness" over time  causes the linking of search results to fail.
Rather than re-indexing all datasets every time a user resets the keying properties
an immutable search key was added to use as a reference.
fixes #364