add ability to pass a datasource in as a query string parameter
pass the name of the datasource as dataSourceName=name and the url
of the datasource as dataSourceUrl=url. if the url is an http
endpoint the data will automatically be fetched. If the endpoint is
accessed via https: a username and password will be requested. Passing
username/password or auth tokens via the URL string will not be supported.